Added code to make train/test date list

+++ mapreduce/unittest/testmkmropt.py
@@ -38,6 +38,10 @@ class TestPreprocess(unittest.TestCase):
         cls.expected_datasets_all3 = os.path.join(cls.data_dir,
                 'expected_datasets_all3.txt')
         cls.test_mktcal = os.path.join(cls.data_dir, 'test_mktcal.csv')
+        cls.expected_datesets = os.path.join(cls.data_dir,
+                'expected_datesets.csv')
+        cls.expected_datesets4 = os.path.join(cls.data_dir,
+                'expected_datesets4.csv')
 
         cls.test_bad_symbol_file = os.path.join(cls.data_dir, 'test_symbols_with_bad.txt')
         cls.expected_bad = os.path.join(cls.data_dir, 'expected_two.csv') # reuse the result
@@ -93,8 +97,9 @@ class TestPreprocess(unittest.TestCase):
                 target.make_symbol_sets(symbols, 2, r_file)
             self.assertTrue(filecmp.cmp(self.expected_datasets_all, r_fn))
         finally:
-            os.remove(r_fn)
-            r_fn = None
+            if r_fn is not None:
+                os.remove(r_fn)
+                r_fn = None
         
     def testDataSets3(self):
         symbols = target.read_symbols(self.test_symbol_file)
@@ -138,21 +143,23 @@ class TestPreprocess(unittest.TestCase):
         first = dt.date(2015, 1, 3)
 
         # invalid train_days
-        self.assertRaises(ValueError, target.extract_dates, cal, first, 0, 1)
+        self.assertRaises(ValueError, target.extract_dates, cal, first, 0, 1, 2)
         # invalid train_inc
-        self.assertRaises(ValueError, target.extract_dates, cal, first, 2, 0)
+        self.assertRaises(ValueError, target.extract_dates, cal, first, 2, 0, 3)
+        # invalid future_day
+        self.assertRaises(ValueError, target.extract_dates, cal, first, 2, 1, 1)
 
     def testExtractDatesNone(self):
         cal = [ dt.date(2015, 1, 2), dt.date(2015, 1, 5), dt.date(2015, 1, 6) ]
 
         # the first date out of range
         first = dt.date(2015, 1, 7)
-        self.assertIsNone(target.extract_dates(cal, first, 2, 1),
+        self.assertIsNone(target.extract_dates(cal, first, 2, 1, 3),
                 'first date out of range')
 
         # the last date out of range
         first - dt.date(2015, 1, 2)
-        self.assertIsNone(target.extract_dates(cal, first, 3, 1),
+        self.assertIsNone(target.extract_dates(cal, first, 3, 1, 4),
                 'last date out of range')
 
     def testExtractDates(self):
@@ -181,25 +188,95 @@ class TestPreprocess(unittest.TestCase):
         first = dt.date(2015, 1, 5)
         days = 10
         inc = 3
+        future = 11
         expected = [ cal[1]
                     ,cal[4]
                     ,cal[7]
                     ,cal[10]
+                    ,cal[1 + future]
                    ]
-        result = target.extract_dates(cal, first, days, inc)
+        result = target.extract_dates(cal, first, days, inc, future)
         self.assertEqual(expected, result, 'On date ' + str(result))
 
         # off date + very last in cal
         first = dt.date(2015, 1, 18)
-        days = 9
+        days = 8
         inc = 2
+        future = 8
         expected = [ cal[11]
                     ,cal[13]
                     ,cal[15]
                     ,cal[17]
-                    ,cal[19]
+                    ,cal[11 + future]
                    ]
-        result = target.extract_dates(cal, first, days, inc)
+        result = target.extract_dates(cal, first, days, inc, future)
         self.assertEqual(expected, result,
                 'Off date + very last in cal' + str(result))
+
+    def testMakeDateSetsException(self):
+        # valid arguments
+        fn_cal = self.test_mktcal
+        refs = [dt.date(2015, 1, 3)]
+        tests = [dt.date(2015, 1, 10)]
+        days = 10
+        inc = 2
+        future = 10
+        with tempfile.TemporaryFile(mode='w') as f_dst:
+            # make sure no exception is raised with valid arguments
+            target.make_date_sets(fn_cal, refs, tests, days, inc, future, f_dst)
+
+            # empty reference
+            self.assertRaises(ValueError, target.make_date_sets,
+                    fn_cal, [], tests, days, inc, future, f_dst)
+            self.assertRaises(ValueError, target.make_date_sets,
+                    fn_cal, None, tests, days, inc, future, f_dst)
+
+            # empty test
+            self.assertRaises(ValueError, target.make_date_sets,
+                    fn_cal, refs, [], days, inc, future, f_dst)
+            self.assertRaises(ValueError, target.make_date_sets,
+                    fn_cal, refs, None, days, inc, future, f_dst)
+
+            # train_inc too large
+            self.assertRaises(ValueError, target.make_date_sets,
+                    fn_cal, refs, tests, days, days, future, f_dst)
+
+            # future_day before the last day of date range
+            self.assertRaises(ValueError, target.make_date_sets,
+                    fn_cal, refs, tests, days, inc, days-1, f_dst)
+
+    def testMakeDateSets(self):
+        fn_cal = self.test_mktcal
+        refs = [dt.date(2015, 1, 3)]
+        tests = [dt.date(2015, 1, 13)]
+        days = 10
+        inc = 2
+        future = 12
+        fn_dst = None
+        try:
+            with tempfile.NamedTemporaryFile('w', prefix='dates_', delete=False) as f_dst:
+                fn_dst = f_dst.name
+                target.make_date_sets(fn_cal, refs, tests, days, inc, future, f_dst)
+            self.assertTrue(filecmp.cmp(self.expected_datesets, fn_dst))
+        finally:
+            pass
+            if fn_dst is not None:
+                os.remove(fn_dst)
+
+    def testMakeDateSets4(self):
+        fn_cal = self.test_mktcal
+        refs = [dt.date(2015, 1, 3), dt.date(2015, 1, 6)]
+        tests = [dt.date(2015, 1, 9), dt.date(2015, 1,11)]
+        days = 8
+        inc = 1
+        future = 10
+        fn_dst = None
+        try:
+            with tempfile.NamedTemporaryFile('w', prefix='dates_', delete=False) as f_dst:
+                fn_dst = f_dst.name
+                target.make_date_sets(fn_cal, refs, tests, days, inc, future, f_dst)
+            self.assertTrue(filecmp.cmp(self.expected_datesets4, fn_dst))
+        finally:
+            if fn_dst is not None:
+                os.remove(fn_dst)
